+++ app/Http/Resources/Config.php
@@ -39,6 +39,7 @@ class Config extends JsonResource
                 'help_url' => $generalSettings->help_url,
                 'legal_notice_url' => $generalSettings->legal_notice_url,
                 'privacy_policy_url' => $generalSettings->privacy_policy_url,
+                'accessibility_statement_url' => $generalSettings->accessibility_statement_url,
                 'no_welcome_page' => $generalSettings->no_welcome_page,
                 'default_locale' => config('app.locale'),
                 'enabled_locales' => array_map(function ($locale) {