+++ lang/en/admin.php
@@ -316,6 +316,10 @@ return [
             'description' => 'If not set, no help button will be displayed.',
             'title' => 'URL to the help page',
         ],
+        'accessibility_statement_url' => [
+            'description' => 'If not set, no accessibility statement link will be displayed in the footer.',
+            'title' => 'URL to the accessibility statement',
+        ],
         'legal_notice_url' => [
             'description' => 'If not set, no legal notice link will be displayed in the footer.',
             'title' => 'URL to the legal notice',